# macrolens_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MacroLens API",
    "description": "宏观数据、发布日历、FOMC、文档检索和 AI 研究平台 API。",
    "version": "1.0.1"
  },
  "paths": {
...
    "/api/v1/series": {
      "get": {
        "tags": [
          "Series"
        ],
        "summary": "List Series",
        "operationId": "list_series_api_v1_series_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "theme",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Theme"
            }
          },
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Provider"
            }
          },
          {
            "name": "frequency",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Frequency"
            }
          },
          {
            "name": "unit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Unit"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "active",
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Series Api V1 Series Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/series/{series_id}": {
      "get": {
        "tags": [
          "Series"
        ],
        "summary": "Series Detail",
        "operationId": "series_detail_api_v1_series__series_id__get",
        "parameters": [
          {
            "name": "series_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Series Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/series/{series_id}/observations": {
      "get": {
        "tags": [
          "Series"
        ],
        "summary": "Series Observations",
        "operationId": "series_observations_api_v1_series__series_id__observations_get",
        "parameters": [
          {
            "name": "series_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Series Id"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          },
          {
            "name": "transform",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(level|difference|mom|qoq|yoy|annualized_3m|annualized_6m|rebased_100|zscore)$",
              "default": "level",
              "title": "Transform"
            }
          },
          {
            "name": "vintage",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "latest",
              "title": "Vintage"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObservationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/series/{series_id}/revisions": {
      "get": {
        "tags": [
          "Series"
        ],
        "summary": "Series Revisions",
        "operationId": "series_revisions_api_v1_series__series_id__revisions_get",
        "parameters": [
          {
            "name": "series_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Series Id"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LicenseInfo": {
        "properties": {
          "display_allowed": {
            "type": "boolean",
            "title": "Display Allowed"
          },
          "download_allowed": {
            "type": "boolean",
            "title": "Download Allowed"
          },
          "api_redistribution_allowed": {
            "type": "boolean",
            "title": "Api Redistribution Allowed"
          },
          "ai_context_allowed": {
            "type": "boolean",
            "title": "Ai Context Allowed"
          },
          "attribution_required": {
            "type": "boolean",
            "title": "Attribution Required"
          },
          "attribution_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attribution Text"
          }
        },
        "type": "object",
        "required": [
          "display_allowed",
          "download_allowed",
          "api_redistribution_allowed",
          "ai_context_allowed",
          "attribution_required"
        ],
        "title": "LicenseInfo"
      },
      "LineageInfo": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "dataset": {
            "type": "string",
            "title": "Dataset"
          },
          "provider_series_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Series Id"
          },
          "source_series_id": {
            "type": "integer",
            "title": "Source Series Id"
          },
          "source_locator": {
            "additionalProperties": true,
            "type": "object",
            "title": "Source Locator"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "dataset",
          "provider_series_id",
          "source_series_id",
          "source_locator"
        ],
        "title": "LineageInfo"
      },
...
      "ObservationMeta": {
        "properties": {
          "data_as_of": {
            "type": "string",
            "format": "date-time",
            "title": "Data As Of"
          },
          "vintage": {
            "type": "string",
            "title": "Vintage"
          },
          "transform": {
            "type": "string",
            "title": "Transform"
          },
          "frequency": {
            "type": "string",
            "title": "Frequency"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "lineage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LineageInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "license": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LicenseInfo"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "data_as_of",
          "vintage",
          "transform",
          "frequency",
          "unit"
        ],
        "title": "ObservationMeta"
      },
      "ObservationPoint": {
        "properties": {
          "period_start": {
            "type": "string",
            "format": "date",
            "title": "Period Start"
          },
          "period_end": {
            "type": "string",
            "format": "date",
            "title": "Period End"
          },
          "value": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "value_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value Text"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "vintage_at": {
            "type": "string",
            "format": "date-time",
            "title": "Vintage At"
          }
        },
        "type": "object",
        "required": [
          "period_start",
          "period_end",
          "value",
          "status",
          "vintage_at"
        ],
        "title": "ObservationPoint"
      },
      "ObservationResponse": {
        "properties": {
          "series": {
            "$ref": "#/components/schemas/SeriesSummary"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/ObservationPoint"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "$ref": "#/components/schemas/ObservationMeta"
          }
        },
        "type": "object",
        "required": [
          "series",
          "data",
          "meta"
        ],
        "title": "ObservationResponse"
      },
...
      "ProviderInfo": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "attribution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attribution"
          },
          "license_class": {
            "type": "string",
            "title": "License Class"
          }
        },
        "type": "object",
        "required": [
          "code",
          "name",
          "license_class"
        ],
        "title": "ProviderInfo"
      },
...
      "RevisionItem": {
        "properties": {
          "period_start": {
            "type": "string",
            "format": "date",
            "title": "Period Start"
          },
          "first_value": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Value"
          },
          "latest_value": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Value"
          },
          "revision": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision"
          },
          "first_vintage_at": {
            "type": "string",
            "format": "date-time",
            "title": "First Vintage At"
          },
          "latest_vintage_at": {
            "type": "string",
            "format": "date-time",
            "title": "Latest Vintage At"
          },
          "versions": {
            "type": "integer",
            "title": "Versions"
          }
        },
        "type": "object",
        "required": [
          "period_start",
          "first_value",
          "latest_value",
          "revision",
          "first_vintage_at",
          "latest_vintage_at",
          "versions"
        ],
        "title": "RevisionItem"
      },
      "RevisionResponse": {
        "properties": {
          "series_id": {
            "type": "string",
            "format": "uuid",
            "title": "Series Id"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/RevisionItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "series_id",
          "items"
        ],
        "title": "RevisionResponse"
      },
...
      "SeriesDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "canonical_code": {
            "type": "string",
            "title": "Canonical Code"
          },
          "name_zh": {
            "type": "string",
            "title": "Name Zh"
          },
          "name_en": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name En"
          },
          "theme": {
            "type": "string",
            "title": "Theme"
          },
          "frequency": {
            "type": "string",
            "title": "Frequency"
          },
          "unit_code": {
            "type": "string",
            "title": "Unit Code"
          },
          "unit_label_zh": {
            "type": "string",
            "title": "Unit Label Zh"
          },
          "default_transform": {
            "type": "string",
            "title": "Default Transform"
          },
          "latest_period": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Period"
          },
          "latest_value": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Value"
          },
          "latest_vintage_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Vintage At"
          },
          "provider": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProviderInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "seasonal_adjustment": {
            "type": "string",
            "title": "Seasonal Adjustment"
          },
          "geography_code": {
            "type": "string",
            "title": "Geography Code"
          },
          "decimal_places": {
            "type": "integer",
            "title": "Decimal Places"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "first_period": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Period"
          },
          "aliases": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Aliases"
          },
          "license": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LicenseInfo"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "canonical_code",
          "name_zh",
          "name_en",
          "theme",
          "frequency",
          "unit_code",
          "unit_label_zh",
          "default_transform",
          "latest_period",
          "description",
          "seasonal_adjustment",
          "geography_code",
          "decimal_places",
          "status",
          "first_period"
        ],
        "title": "SeriesDetail"
      },
      "SeriesSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "canonical_code": {
            "type": "string",
            "title": "Canonical Code"
          },
          "name_zh": {
            "type": "string",
            "title": "Name Zh"
          },
          "name_en": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name En"
          },
          "theme": {
            "type": "string",
            "title": "Theme"
          },
          "frequency": {
            "type": "string",
            "title": "Frequency"
          },
          "unit_code": {
            "type": "string",
            "title": "Unit Code"
          },
          "unit_label_zh": {
            "type": "string",
            "title": "Unit Label Zh"
          },
          "default_transform": {
            "type": "string",
            "title": "Default Transform"
          },
          "latest_period": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Period"
          },
          "latest_value": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Value"
          },
          "latest_vintage_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Vintage At"
          },
          "provider": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProviderInfo"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "canonical_code",
          "name_zh",
          "name_en",
          "theme",
          "frequency",
          "unit_code",
          "unit_label_zh",
          "default_transform",
          "latest_period"
        ],
        "title": "SeriesSummary"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}